fix: replace cli-optique to cli - replace console.log to print, console.error to printError from optique - remove unused packages
