--- packages/cli-optique/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@fedify/cli-optique",
-  "version": "2.0.0",
-  "private": true,
-  "description": "CLI toolchain for Fedify and debugging ActivityPub",
-  "keywords": [
-    "fedify",
-    "activitypub",
-    "cli",
-    "fediverse"
-  ],
-  "homepage": "https://fedify.dev/cli",
-  "bugs": {
-    "url": "https://github.com/fedify-dev/fedify/issues"
-  },
-  "license": "MIT",
-  "author": {
-    "name": "Hong Minhee",
-    "email": "hong@minhee.org",
-    "url": "https://hongminhee.org/"
-  },
-  "funding": [
-    "https://opencollective.com/fedify",
-    "https://github.com/sponsors/dahlia"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/fedify-dev/fedify.git",
-    "directory": "packages/cli"
-  },
-  "type": "module",
-  "engines": {
-    "node": ">=20.0.0",
-    "bun": ">=1.2.0",
-    "denp": ">=2.0.0"
-  },
-  "dependencies": {
-    "@fedify/fedify": "workspace:",
-    "@optique/core": "^0.3.0",
-    "@optique/run": "^0.3.0",
-    "cli-highlight": "^2.1.11",
-    "icojs": "^0.19.5",
-    "jimp": "^1.6.0",
-    "@jimp/core": "^1.6.0",
-    "@jimp/wasm-webp": "^1.6.0",
-    "ora": "^8.2.0",
-    "@cross/dir": "jsr:@cross/dir@^1.1.0",
-    "cli-highlight": "^2.1.11",
-    "@hongminhee/localtunnel": "jsr:@hongminhee/localtunnel@^0.2.0",
-    "@std/fmt/colors": "jsr:@std/fmt@^0.224.0/colors",
-    "@cliffy/command": "jsr:@cliffy/command@1.0.0-rc.4"
-  }
-}
