--- packages/cli/.gitignore
@@ -1,4 +1,0 @@
-bin/
-fedify-cli-*.tar.xz
-fedify-cli-*.tgz
-fedify-cli-*.zip
