+++ packages/cli/deno.json
@@ -4,39 +4,16 @@
   "license": "MIT",
   "exports": "./src/mod.ts",
   "imports": {
-    "@cliffy/ansi": "jsr:@cliffy/ansi@1.0.0-rc.4",
-    "@cliffy/command": "jsr:@cliffy/command@1.0.0-rc.4",
-    "@cliffy/prompt": "jsr:@cliffy/prompt@1.0.0-rc.4",
-    "@cliffy/table": "jsr:@cliffy/table@1.0.0-rc.4",
     "@cross/dir": "jsr:@cross/dir@^1.1.0",
-    "@david/dax": "jsr:@david/dax@^0.41.0",
+    "@optique/core": "jsr:@optique/core@^0.3.0",
+    "@optique/run": "jsr:@optique/run@^0.3.0",
+    "@std/fmt": "jsr:@std/fmt@^1.0.8",
+    "ora": "npm:ora@^8.2.0",
     "@hongminhee/localtunnel": "jsr:@hongminhee/localtunnel@^0.2.0",
-    "@jimp/core": "npm:@jimp/core@^1.6.0",
     "@jimp/wasm-webp": "npm:@jimp/wasm-webp@^1.6.0",
-    "@poppanator/http-constants": "npm:@poppanator/http-constants@^1.1.1",
-    "@std/assert": "jsr:@std/assert@^1.0.13",
-    "@std/fmt/colors": "jsr:@std/fmt@^0.224.0/colors",
-    "@std/testing": "jsr:@std/testing@^1.0.8",
-    "@std/dotenv": "jsr:@std/dotenv@^0.225.2",
-    "@std/semver": "jsr:@std/semver@^1.0.5",
-    "cli-highlight": "npm:cli-highlight@^2.1.11",
-    "fetch-mock": "npm:fetch-mock@^12.5.2",
-    "hono": "jsr:@hono/hono@^4.8.3",
-    "icojs": "npm:icojs@^0.19.4",
-    "jimp": "npm:jimp@^1.6.0",
-    "ora": "npm:ora@^8.0.1",
-    "shiki": "npm:shiki@^1.6.4",
-    "sharp": "npm:sharp@^0.34.3"
+    "@jimp/core": "npm:@jimp/core@^1.6.0",
+    "@std/fmt/colors": "jsr:@std/fmt@^0.224.0/colors"
   },
-  "exclude": [
-    ".vscode",
-    "fedify-cli-*.tar.xz",
-    "fedify-cli-*.tgz",
-    "fedify-cli-*.zip",
-    "package.json",
-    "src/install.mjs",
-    "src/run.mjs"
-  ],
   "tasks": {
     "codegen": "deno task -f @fedify/fedify codegen",
     "check": {
@@ -51,18 +28,6 @@
         "codegen"
       ]
     },
-    "publish": {
-      "command": "deno publish",
-      "dependencies": [
-        "codegen"
-      ]
-    },
-    "publish-dry-run": "deno task publish --dry-run --allow-dirty",
-    "pack": {
-      "command": "deno run -A scripts/pack.ts",
-      "dependencies": [
-        "codegen"
-      ]
-    }
+    "runi": "deno run --allow-all src/mod.ts"
   }
 }
