+++ packages/cli/package.json
@@ -1,31 +1,7 @@
 {
   "name": "@fedify/cli",
   "version": "2.0.0",
-  "type": "module",
-  "files": [
-    "README.md",
-    "package.json",
-    "src/install.mjs",
-    "src/run.mjs"
-  ],
-  "bin": {
-    "fedify": "./src/run.mjs"
-  },
-  "scripts": {
-    "postinstall": "node src/install.mjs"
-  },
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "os": [
-    "darwin",
-    "linux",
-    "win32"
-  ],
-  "cpu": [
-    "x64",
-    "arm64"
-  ],
+  "private": true,
   "description": "CLI toolchain for Fedify and debugging ActivityPub",
   "keywords": [
     "fedify",
@@ -51,5 +27,25 @@
     "type": "git",
     "url": "git+https://github.com/fedify-dev/fedify.git",
     "directory": "packages/cli"
+  },
+  "type": "module",
+  "engines": {
+    "node": ">=20.0.0",
+    "bun": ">=1.2.0",
+    "denp": ">=2.0.0"
+  },
+  "dependencies": {
+    "@fedify/fedify": "workspace:",
+    "@optique/core": "^0.3.0",
+    "@optique/run": "^0.3.0",
+    "cli-highlight": "^2.1.11",
+    "jimp": "^1.6.0",
+    "@jimp/core": "^1.6.0",
+    "@jimp/wasm-webp": "^1.6.0",
+    "ora": "^8.2.0",
+    "@cross/dir": "jsr:@cross/dir@^1.1.0",
+    "cli-highlight": "^2.1.11",
+    "@hongminhee/localtunnel": "jsr:@hongminhee/localtunnel@^0.2.0",
+    "@std/fmt/colors": "jsr:@std/fmt@^0.224.0/colors"
   }
 }
