+++ packages/cli/src/lookup.ts
@@ -1,4 +1,22 @@
-import { Command, EnumType } from "@cliffy/command";
+import {
+  argument,
+  choice,
+  constant,
+  flag,
+  float,
+  type InferValue,
+  map,
+  merge,
+  message,
+  multiple,
+  object,
+  option,
+  optional,
+  or,
+  string,
+  withDefault,
+} from "@optique/core";
+import { path } from "@optique/run/valueparser";
 import {
   Application,
   Collection,
@@ -14,75 +32,122 @@ import {
   traverseCollection,
 } from "@fedify/fedify";
 import { getLogger } from "@logtape/logtape";
-import * as colors from "@std/fmt/colors";
-import { dirname, isAbsolute, resolve } from "@std/path";
 import ora from "ora";
+import * as colors from "@std/fmt/colors";
+import process from "node:process";
+import { createWriteStream, type WriteStream } from "node:fs";
 import { getContextLoader, getDocumentLoader } from "./docloader.ts";
-import { renderImages } from "./imagerenderer.ts";
 import { spawnTemporaryServer, type TemporaryServer } from "./tempserver.ts";
 import { colorEnabled, formatObject } from "./utils.ts";
+import { renderImages } from "./imagerenderer.ts";
+import { print, printError } from "@optique/run";
 
 const logger = getLogger(["fedify", "cli", "lookup"]);
 
-const sigSpec = new EnumType(["draft-cavage-http-signatures-12", "rfc9421"]);
+const authorizedFetchOption = withDefault(
+  object({
+    authorizedFetch: flag("-a", "--authorized-fetch", {
+      description: message`Sign the request with an one-time key.`,
+    }),
+    firstKnock: withDefault(
+      option(
+        "--first-knock",
+        choice(["draft-cavage-http-signatures-12", "rfc9421"]),
+        {
+          description:
+            message`The first-knock spec for -a/--authorized-fetch. It is used for the double-knocking technique.`,
+        },
+      ),
+      "draft-cavage-http-signatures-12" as const,
+    ),
+  }),
+  { authorizedFetch: false } as const,
+);
 
-interface CommandOptions {
-  authorizedFetch?: boolean;
-  firstKnock: "draft-cavage-http-signatures-12" | "rfc9421";
-  traverse?: boolean;
-  suppressErrors?: boolean;
-  raw?: boolean;
-  compact?: boolean;
-  expand?: boolean;
-  userAgent?: string;
-  separator: string;
-  output?: string;
-  timeout?: number;
-}
+const traverseOption = withDefault(
+  object({
+    traverse: flag("-t", "--traverse", {
+      description:
+        message`Traverse the given collection to fetch all items. If it is turned on, the argument cannot be multiple.`,
+    }),
+    suppressErrors: option("-S", "--suppress-errors", {
+      description:
+        message`Suppress partial errors while traversing the collection`,
+    }),
+  }),
+  { traverse: false } as const,
+);
 
-export async function createFileStream(
-  outputPath: string,
-): Promise<WritableStream> {
-  try {
-    const filepath = isAbsolute(outputPath)
-      ? outputPath
-      : resolve(Deno.env.get("PWD") || Deno.cwd(), outputPath);
+export const lookupCommand = merge(
+  traverseOption,
+  authorizedFetchOption,
+  object({
+    command: constant("lookup"),
+    format: withDefault(
+      or(
+        map(
+          option("-r", "--raw", {
+            description: message`Print the fetched JSON-LD document as is.`,
+          }),
+          () => "raw" as const,
+        ),
+        map(
+          option("-C", "--compact", {
+            description: message`Compact the fetched JSON-LD document.`,
+          }),
+          () => "compact" as const,
+        ),
+        map(
+          option("-e", "--expand", {
+            description: message`Expand the fetched JSON-LD document.`,
+          }),
+          () => "expand" as const,
+        ),
+      ),
+      "default" as const,
+    ),
+    userAgent: optional(
+      option("-u", "--user-agent", string({ metavar: "USER_AGENT" }), {
+        description: message`The custom User-Agent header value.`,
+      }),
+    ),
+    separator: withDefault(
+      option("-s", "--separator", string({ metavar: "SEPARATOR" }), {
+        description:
+          message`Specify the separator between adjacent output objects or collection items.`,
+      }),
+      "----",
+    ),
+    output: optional(option(
+      "-o",
+      "--output",
+      path({
+        metavar: "OUTPUT_PATH",
+        type: "file",
+        allowCreate: true,
+      }),
+      { description: message`Specify the output file path.` },
+    )),
+    timeout: optional(option(
+      "-T",
+      "--timeout",
+      float({ min: 0, metavar: "SECONDS" }),
+      { description: message`Set timeout for network requests in seconds.` },
+    )),
+    urls: multiple(
+      argument(string({ metavar: "URL_OR_HANDLE" }), {
+        description: message`One or more URLs or handles to look up.`,
+      }),
+      { min: 1 },
+    ),
+  }),
+);
 
-    const parentDir = dirname(filepath);
-    await Deno.mkdir(parentDir, { recursive: true });
-
-    const file = await Deno.open(filepath, {
-      write: true,
-      create: true,
-      truncate: true,
-    });
-
-    return new WritableStream({
-      write: (chunk) => file.write(chunk).then(() => {}),
-      close: () => file.close(),
-      abort: (reason) => {
-        file.close();
-        throw reason;
-      },
-    });
-  } catch (err) {
-    const spinner = ora({
-      text: `Failed to write output to ${colors.red(outputPath)}.`,
-      discardStdin: false,
-    });
-    spinner.fail();
-    console.error(`Error: ${String(err)}`);
-
-    if (err instanceof Deno.errors.PermissionDenied) {
-      console.error(
-        "Permission denied. Try running with proper permissions.",
-      );
-    } else if (err instanceof Deno.errors.NotFound) {
-      console.error("Path does not exist or is invalid.");
-    } else if (err instanceof Deno.errors.IsADirectory) {
-      console.error("The specified path is a directory, not a file.");
-    }
-    Deno.exit(1);
+export class TimeoutError extends Error {
+  override name = "TimeoutError";
+  constructor(message: string) {
+    super(message);
+    this.name = "TimeoutError";
   }
 }
 
@@ -103,50 +168,47 @@ async function findAllImages(obj: APObject): Promise<URL[]> {
 
 export async function writeObjectToStream(
   object: APObject | Link,
-  options: CommandOptions,
+  outputPath: string | undefined,
+  format: string | undefined,
   contextLoader: DocumentLoader,
 ): Promise<void> {
-  const stream = options.output
-    ? await createFileStream(options.output)
-    : Deno.stdout.writable;
+  const stream: WriteStream | NodeJS.WritableStream = outputPath
+    ? createWriteStream(outputPath)
+    : process.stdout;
 
-  const writer = stream.getWriter();
+  let content;
+  let json = true;
+  let imageUrls: URL[] = [];
 
-  try {
-    let content;
-    let json = true;
-    let imageUrls: URL[] = [];
-
-    if (options.raw) {
+  if (format) {
+    if (format === "raw") {
       content = await object.toJsonLd({ contextLoader });
-    } else if (options.compact) {
+    } else if (format === "compact") {
       content = await object.toJsonLd({ format: "compact", contextLoader });
-    } else if (options.expand) {
+    } else if (format === "expand") {
       content = await object.toJsonLd({ format: "expand", contextLoader });
     } else {
       content = object;
       json = false;
     }
+  } else {
+    content = object;
+    json = false;
+  }
 
-    const enableColors = colorEnabled && options.output === undefined;
-    content = formatObject(content, enableColors, json);
+  const enableColors = colorEnabled && outputPath === undefined;
+  content = formatObject(content, enableColors, json);
 
-    const encoder = new TextEncoder();
-    const bytes = encoder.encode(content + "\n");
+  const encoder = new TextEncoder();
+  const bytes = encoder.encode(content + "\n");
 
-    await writer.write(bytes);
+  stream.write(bytes);
 
-    if (object instanceof APObject) {
-      imageUrls = await findAllImages(object);
-    }
-    if (!options.output && imageUrls.length > 0) {
-      await renderImages(imageUrls);
-    }
-  } finally {
-    writer.releaseLock();
-    if (options.output) {
-      await stream.close();
-    }
+  if (object instanceof APObject) {
+    imageUrls = await findAllImages(object);
+  }
+  if (!outputPath && imageUrls.length > 0) {
+    await renderImages(imageUrls);
   }
 }
 
@@ -159,9 +221,7 @@ export function createTimeoutSignal(
   const controller = new AbortController();
   const timerId = setTimeout(() => {
     controller.abort(
-      new Deno.errors.TimedOut(
-        `Request timed out after ${timeoutSeconds} seconds`,
-      ),
+      new TimeoutError(`Request timed out after ${timeoutSeconds} seconds`),
     );
   }, timeoutSeconds * 1000);
 
@@ -179,18 +239,6 @@ export function clearTimeoutSignal(signal?: AbortSignal): void {
   }
 }
 
-function handleTimeoutError(
-  spinner: { fail: (text: string) => void },
-  timeoutSeconds?: number,
-  url?: string,
-): void {
-  const urlText = url ? ` for: ${colors.red(url)}` : "";
-  spinner.fail(`Request timed out after ${timeoutSeconds} seconds${urlText}.`);
-  console.error(
-    "Try increasing the timeout with -T/--timeout option or check network connectivity.",
-  );
-}
-
 function wrapDocumentLoaderWithTimeout(
   loader: DocumentLoader,
   timeoutSeconds?: number,
@@ -205,299 +253,279 @@ function wrapDocumentLoaderWithTimeout(
   };
 }
 
-export const command = new Command()
-  .type("sig-spec", sigSpec)
-  .arguments("<...urls:string>")
-  .description(
-    "Lookup an Activity Streams object by URL or the actor handle.  " +
-      "The argument can be either a URL or an actor handle " +
-      "(e.g., @username@domain), and it can be multiple.",
-  )
-  .option("-a, --authorized-fetch", "Sign the request with an one-time key.")
-  .option(
-    "--first-knock <spec:sig-spec>",
-    "The first-knock spec for -a/--authorized-fetch.  It is used for " +
-      "the double-knocking technique.",
-    { depends: ["authorized-fetch"], default: "rfc9421" },
-  )
-  .option(
-    "-t, --traverse",
-    "Traverse the given collection to fetch all items.  If it is turned on, " +
-      "the argument cannot be multiple.",
-  )
-  .option(
-    "-S, --suppress-errors",
-    "Suppress partial errors while traversing the collection.",
-    { depends: ["traverse"] },
-  )
-  .option("-r, --raw", "Print the fetched JSON-LD document as is.", {
-    conflicts: ["compact", "expand"],
-  })
-  .option("-C, --compact", "Compact the fetched JSON-LD document.", {
-    conflicts: ["raw", "expand"],
-  })
-  .option("-e, --expand", "Expand the fetched JSON-LD document.", {
-    conflicts: ["raw", "compact"],
-  })
-  .option("-u, --user-agent <string>", "The custom User-Agent header value.")
-  .option(
-    "-s, --separator <string>",
-    "Specify the separator between adjacent output objects or " +
-      "collection items.",
-    { default: "----" },
-  )
-  .option(
-    "-o, --output <file>",
-    "Specify the output file path.",
-  )
-  .option(
-    "-T, --timeout <seconds:number>",
-    "Set timeout for network requests in seconds.",
-  )
-  .action(async (options, ...urls: string[]) => {
-    if (urls.length < 1) {
-      console.error("At least one URL or actor handle must be provided.");
-      Deno.exit(1);
-    } else if (options.traverse && urls.length > 1) {
-      console.error(
-        "The -t/--traverse option cannot be used with multiple arguments.",
-      );
-      Deno.exit(1);
-    }
+function handleTimeoutError(
+  spinner: { fail: (text: string) => void },
+  timeoutSeconds?: number,
+  url?: string,
+): void {
+  const urlText = url ? ` for: ${colors.red(url)}` : "";
+  spinner.fail(`Request timed out after ${timeoutSeconds} seconds${urlText}.`);
+  printError(
+    message`Try increasing the timeout with -T/--timeout option or check network connectivity.`,
+  );
+}
 
-    const spinner = ora({
-      text: `Looking up the ${
-        options.traverse ? "collection" : urls.length > 1 ? "objects" : "object"
-      }...`,
-      discardStdin: false,
-    }).start();
-    let server: TemporaryServer | undefined = undefined;
-    const baseDocumentLoader = await getDocumentLoader({
-      userAgent: options.userAgent,
-    });
-    const documentLoader = wrapDocumentLoaderWithTimeout(
-      baseDocumentLoader,
-      options.timeout,
+export async function runLookup(command: InferValue<typeof lookupCommand>) {
+  // FIXME: Implement -t, --traverse when multiple URLs are provided
+  if (command.urls.length < 1) {
+    printError(message`At least one URL or actor handle must be provided.`);
+    process.exit(1);
+  } else if (command.traverse && command.urls.length > 1) {
+    printError(
+      message`The -t/--traverse option cannot be used with multiple arguments.`,
     );
-    const baseContextLoader = await getContextLoader({
-      userAgent: options.userAgent,
-    });
-    const contextLoader = wrapDocumentLoaderWithTimeout(
-      baseContextLoader,
-      options.timeout,
-    );
-    let authLoader: DocumentLoader | undefined = undefined;
-    if (options.authorizedFetch) {
-      spinner.text = "Generating a one-time key pair...";
-      const key = await generateCryptoKeyPair();
-      spinner.text = "Spinning up a temporary ActivityPub server...";
-      server = await spawnTemporaryServer((req) => {
-        const serverUrl = server?.url ?? new URL("http://localhost/");
-        if (new URL(req.url).pathname == "/.well-known/webfinger") {
-          const jrd: ResourceDescriptor = {
-            subject: `acct:${serverUrl.hostname}@${serverUrl.hostname}`,
-            aliases: [serverUrl.href],
-            links: [
-              {
-                rel: "self",
-                href: serverUrl.href,
-                type: "application/activity+json",
-              },
-            ],
-          };
-          return new Response(JSON.stringify(jrd), {
-            headers: { "Content-Type": "application/jrd+json" },
-          });
-        }
-        return respondWithObject(
-          new Application({
-            id: serverUrl,
-            preferredUsername: serverUrl?.hostname,
-            publicKey: new CryptographicKey({
-              id: new URL("#main-key", serverUrl),
-              owner: serverUrl,
-              publicKey: key.publicKey,
-            }),
-            manuallyApprovesFollowers: true,
-            inbox: new URL("/inbox", serverUrl),
-            outbox: new URL("/outbox", serverUrl),
+    process.exit(1);
+  }
+  const spinner = ora({
+    text: `Looking up the ${
+      command.traverse
+        ? "collection"
+        : command.urls.length > 1
+        ? "objects"
+        : "object"
+    }...`,
+    discardStdin: false,
+  }).start();
+
+  let server: TemporaryServer | undefined = undefined;
+  const baseDocumentLoader = await getDocumentLoader({
+    userAgent: command.userAgent,
+  });
+  const documentLoader = wrapDocumentLoaderWithTimeout(
+    baseDocumentLoader,
+    command.timeout,
+  );
+  const baseContextLoader = await getContextLoader({
+    userAgent: command.userAgent,
+  });
+  const contextLoader = wrapDocumentLoaderWithTimeout(
+    baseContextLoader,
+    command.timeout,
+  );
+
+  let authLoader: DocumentLoader | undefined = undefined;
+
+  if (command.authorizedFetch) {
+    spinner.text = "Generating a one-time key pair...";
+    const key = await generateCryptoKeyPair();
+    spinner.text = "Spinning up a temporary ActivityPub server...";
+    server = await spawnTemporaryServer((req) => {
+      const serverUrl = server?.url ?? new URL("http://localhost/");
+      if (new URL(req.url).pathname == "/.well-known/webfinger") {
+        const jrd: ResourceDescriptor = {
+          subject: `acct:${serverUrl.hostname}@${serverUrl.hostname}`,
+          aliases: [serverUrl.href],
+          links: [
+            {
+              rel: "self",
+              href: serverUrl.href,
+              type: "application/activity+json",
+            },
+          ],
+        };
+        return new Response(JSON.stringify(jrd), {
+          headers: { "Content-Type": "application/jrd+json" },
+        });
+      }
+      return respondWithObject(
+        new Application({
+          id: serverUrl,
+          preferredUsername: serverUrl?.hostname,
+          publicKey: new CryptographicKey({
+            id: new URL("#main-key", serverUrl),
+            owner: serverUrl,
+            publicKey: key.publicKey,
           }),
-          { contextLoader },
-        );
-      });
-      const baseAuthLoader = getAuthenticatedDocumentLoader(
-        {
-          keyId: new URL("#main-key", server.url),
-          privateKey: key.privateKey,
-        },
-        {
-          specDeterminer: {
-            determineSpec() {
-              return options.firstKnock;
-            },
-            rememberSpec() {
-            },
+          manuallyApprovesFollowers: true,
+          inbox: new URL("/inbox", serverUrl),
+          outbox: new URL("/outbox", serverUrl),
+        }),
+        { contextLoader },
+      );
+    });
+    const baseAuthLoader = getAuthenticatedDocumentLoader(
+      {
+        keyId: new URL("#main-key", server.url),
+        privateKey: key.privateKey,
+      },
+      {
+        specDeterminer: {
+          determineSpec() {
+            return command.firstKnock;
+          },
+          rememberSpec() {
           },
         },
-      );
-      authLoader = wrapDocumentLoaderWithTimeout(
-        baseAuthLoader,
-        options.timeout,
-      );
-    }
+      },
+    );
+    authLoader = wrapDocumentLoaderWithTimeout(
+      baseAuthLoader,
+      command.timeout,
+    );
+  }
 
-    spinner.text = `Looking up the ${
-      options.traverse ? "collection" : urls.length > 1 ? "objects" : "object"
-    }...`;
+  spinner.text = `Looking up the ${
+    command.traverse
+      ? "collection"
+      : command.urls.length > 1
+      ? "objects"
+      : "object"
+  }...`;
 
-    if (options.traverse) {
-      const url = urls[0];
-      let collection: APObject | null;
-      try {
-        collection = await lookupObject(url, {
-          documentLoader: authLoader ?? documentLoader,
-          contextLoader,
-          userAgent: options.userAgent,
-        });
-      } catch (error) {
-        if (error instanceof Deno.errors.TimedOut) {
-          handleTimeoutError(spinner, options.timeout, url);
-        } else {
-          spinner.fail(`Failed to fetch object: ${colors.red(url)}.`);
-          if (authLoader == null) {
-            console.error(
-              "It may be a private object.  Try with -a/--authorized-fetch.",
-            );
-          }
-        }
-        await server?.close();
-        Deno.exit(1);
-      }
-      if (collection == null) {
+  if (command.traverse) {
+    const url = command.urls[0];
+    let collection: APObject | null;
+    try {
+      collection = await lookupObject(url, {
+        documentLoader: authLoader ?? documentLoader,
+        contextLoader,
+        userAgent: command.userAgent,
+      });
+    } catch (error) {
+      if (error instanceof TimeoutError) {
+        handleTimeoutError(spinner, command.timeout, url);
+      } else {
         spinner.fail(`Failed to fetch object: ${colors.red(url)}.`);
         if (authLoader == null) {
-          console.error(
-            "It may be a private object.  Try with -a/--authorized-fetch.",
+          printError(
+            message`It may be a private object.  Try with -a/--authorized-fetch.`,
           );
         }
-        await server?.close();
-        Deno.exit(1);
       }
-      if (!(collection instanceof Collection)) {
-        spinner.fail(
-          `Not a collection: ${colors.red(url)}.  ` +
-            "The -t/--traverse option requires a collection.",
+      await server?.close();
+      process.exit(1);
+    }
+    if (collection == null) {
+      spinner.fail(`Failed to fetch object: ${colors.red(url)}.`);
+      if (authLoader == null) {
+        printError(
+          message`It may be a private object.  Try with -a/--authorized-fetch.`,
         );
-        await server?.close();
-        Deno.exit(1);
       }
-      spinner.succeed(`Fetched collection: ${colors.green(url)}.`);
-      try {
-        let i = 0;
-        for await (
-          const item of traverseCollection(collection, {
-            documentLoader: authLoader ?? documentLoader,
-            contextLoader,
-            suppressError: options.suppressErrors,
-          })
-        ) {
-          if (!options.output && i > 0) console.log(options.separator);
-          await writeObjectToStream(item, options, contextLoader);
-          i++;
-        }
-      } catch (error) {
-        logger.error("Failed to complete the traversal: {error}", { error });
-        if (error instanceof Deno.errors.TimedOut) {
-          handleTimeoutError(spinner, options.timeout);
-        } else {
-          spinner.fail("Failed to complete the traversal.");
-          if (authLoader == null) {
-            console.error(
-              "It may be a private object.  Try with -a/--authorized-fetch.",
-            );
-          } else {
-            console.error(
-              "Use the -S/--suppress-errors option to suppress partial errors.",
-            );
-          }
-        }
-        await server?.close();
-        Deno.exit(1);
-      }
-      spinner.succeed("Successfully fetched all items in the collection.");
-
       await server?.close();
-      Deno.exit(0);
+      process.exit(1);
     }
-
-    const promises: Promise<APObject | null>[] = [];
-    for (const url of urls) {
-      promises.push(
-        lookupObject(
-          url,
-          {
-            documentLoader: authLoader ?? documentLoader,
-            contextLoader,
-            userAgent: options.userAgent,
-          },
-        ).catch((error) => {
-          if (error instanceof Deno.errors.TimedOut) {
-            handleTimeoutError(spinner, options.timeout, url);
-          }
-          throw error;
-        }),
+    if (!(collection instanceof Collection)) {
+      spinner.fail(
+        `Not a collection: ${colors.red(url)}.  ` +
+          "The -t/--traverse option requires a collection.",
       );
+      await server?.close();
+      process.exit(1);
     }
+    spinner.succeed(`Fetched collection: ${colors.green(url)}.`);
 
-    let objects: (APObject | null)[];
     try {
-      objects = await Promise.all(promises);
-    } catch (_error) {
-      await server?.close();
-      Deno.exit(1);
-    }
-    let success = true;
-    let i = 0;
-    for (const object of objects) {
-      const url = urls[i];
-      if (i > 0) console.log(options.separator);
-      i++;
-      try {
-        if (object == null) {
-          spinner.fail(`Failed to fetch object: ${colors.red(url)}.`);
-          if (authLoader == null) {
-            console.error(
-              "It may be a private object.  Try with -a/--authorized-fetch.",
-            );
-          }
-          success = false;
+      let i = 0;
+      for await (
+        const item of traverseCollection(collection, {
+          documentLoader: authLoader ?? documentLoader,
+          contextLoader,
+          suppressError: command.suppressErrors,
+        })
+      ) {
+        if (!command.output && i > 0) print(message`${command.separator}`);
+        await writeObjectToStream(
+          item,
+          command.output,
+          command.format,
+          contextLoader,
+        );
+        i++;
+      }
+    } catch (error) {
+      logger.error("Failed to complete the traversal: {error}", { error });
+      if (error instanceof TimeoutError) {
+        handleTimeoutError(spinner, command.timeout);
+      } else {
+        spinner.fail("Failed to complete the traversal.");
+        if (authLoader == null) {
+          printError(
+            message`It may be a private object.  Try with -a/--authorized-fetch.`,
+          );
         } else {
-          spinner.succeed(`Fetched object: ${colors.green(url)}.`);
-          await writeObjectToStream(object, options, contextLoader);
-          if (i < urls.length - 1) {
-            console.log(options.separator);
-          }
+          printError(
+            message`Use the -S/--suppress-errors option to suppress partial errors.`,
+          );
         }
-      } catch (_) {
-        success = false;
+      }
+      await server?.close();
+      process.exit(1);
+    }
+    spinner.succeed("Successfully fetched all items in the collection.");
+
+    await server?.close();
+    process.exit(0);
+  }
+
+  const promises: Promise<APObject | null>[] = [];
+
+  for (const url of command.urls) {
+    promises.push(
+      lookupObject(url, {
+        documentLoader: authLoader ?? documentLoader,
+        contextLoader,
+        userAgent: command.userAgent,
+      }).catch((error) => {
+        if (error instanceof TimeoutError) {
+          handleTimeoutError(spinner, command.timeout, url);
+        }
+        throw error;
+      }),
+    );
+  }
+
+  let objects: (APObject | null)[];
+  try {
+    objects = await Promise.all(promises);
+  } catch (_error) {
+    //TODO: implement -a --authorized-fetch
+    // await server?.close();
+    process.exit(1);
+  }
+
+  spinner.stop();
+  let success = true;
+  let i = 0;
+  for (const obj of objects) {
+    const url = command.urls[i];
+    if (i > 0) print(message`${command.separator}`);
+    i++;
+    if (obj == null) {
+      spinner.fail(`Failed to fetch ${colors.red(url)}`);
+      if (authLoader == null) {
+        printError(
+          message`It may be a private object.  Try with -a/--authorized-fetch.`,
+        );
+      }
+      success = false;
+    } else {
+      spinner.succeed(`Fetched object: ${colors.green(url)}`);
+      await writeObjectToStream(
+        obj,
+        command.output,
+        command.format,
+        contextLoader,
+      );
+      if (i < command.urls.length - 1) {
+        print(message`${command.separator}`);
       }
     }
-    if (success) {
-      spinner.succeed(
-        urls.length > 1
-          ? "Successfully fetched all objects."
-          : "Successfully fetched the object.",
-      );
-    }
-    await server?.close();
-    if (!success) {
-      Deno.exit(1);
-    }
-    if (success && options.output) {
-      spinner.succeed(
-        `Successfully wrote output to ${colors.green(options.output)}.`,
-      );
-    }
-  });
-
-// cSpell: ignore sigspec
+  }
+  if (success) {
+    spinner.succeed(
+      command.urls.length > 1
+        ? "Successfully fetched all objects."
+        : "Successfully fetched the object.",
+    );
+  }
+  await server?.close();
+  if (!success) {
+    process.exit(1);
+  }
+  if (success && command.output) {
+    spinner.succeed(
+      `Successfully wrote output to ${colors.green(command.output)}.`,
+    );
+  }
+}
