+++ packages/cli/src/mod.ts
@@ -1,79 +1,44 @@
-import { Command, CompletionsCommand, HelpCommand } from "@cliffy/command";
-import { getFileSink } from "@logtape/file";
-import { configure, getConsoleSink } from "@logtape/logtape";
-import { setColorEnabled } from "@std/fmt/colors";
-import { AsyncLocalStorage } from "node:async_hooks";
-import { DEFAULT_CACHE_DIR, setCacheDir } from "./cache.ts";
-import metadata from "../deno.json" with { type: "json" };
-import { command as inbox } from "./inbox.tsx";
-import { command as init } from "./init.ts";
-import { logFile, recordingSink } from "./log.ts";
-import { command as lookup } from "./lookup.ts";
-import { command as nodeinfo } from "./nodeinfo.ts";
-import { command as tunnel } from "./tunnel.ts";
-import { colorEnabled } from "./utils.ts";
-import { command as webfinger } from "./webfinger.ts";
+import { or } from "@optique/core";
+import { run } from "@optique/run";
+import { lookupCommand, runLookup } from "./lookup.ts";
+import { runWebFinger, webFingerCommand } from "./webfinger.ts";
+import { initCommand, runInit } from "./init.ts";
+import { inboxCommand, runInbox } from "./inbox.ts";
+import { nodeInfoCommand, runNodeInfo } from "./nodeinfo.ts";
+import { runTunnel, tunnelCommand } from "./tunnel.ts";
 
-setColorEnabled(colorEnabled);
+const command = or(
+  initCommand,
+  lookupCommand,
+  inboxCommand,
+  nodeInfoCommand,
+  tunnelCommand,
+  webFingerCommand,
+);
 
-const command = new Command()
-  .name("fedify")
-  .version(metadata.version)
-  .help({ colors: colorEnabled })
-  .globalEnv(
-    "FEDIFY_LOG_FILE=<file:file>",
-    "An optional file to write logs to.  " +
-      "Regardless of -d/--debug option, " +
-      "all levels of logs are written to this file.  " +
-      "Note that this does not mute console logs.",
-  )
-  .globalOption("-d, --debug", "Enable debug mode.", {
-    async action() {
-      await configure({
-        sinks: {
-          console: getConsoleSink(),
-          recording: recordingSink,
-          file: logFile == null ? () => undefined : getFileSink(logFile),
-        },
-        filters: {},
-        loggers: [
-          {
-            category: "fedify",
-            lowestLevel: "debug",
-            sinks: ["console", "recording", "file"],
-          },
-          {
-            category: "localtunnel",
-            lowestLevel: "debug",
-            sinks: ["console", "file"],
-          },
-          {
-            category: ["logtape", "meta"],
-            lowestLevel: "warning",
-            sinks: ["console", "file"],
-          },
-        ],
-        reset: true,
-        contextLocalStorage: new AsyncLocalStorage(),
-      });
-    },
-  })
-  .globalOption("-c, --cache-dir=<dir:file>", "Set the cache directory.", {
-    default: DEFAULT_CACHE_DIR,
-    async action(options) {
-      await setCacheDir(options.cacheDir);
-    },
-  })
-  .default("help")
-  .command("init", init)
-  .command("lookup", lookup)
-  .command("inbox", inbox)
-  .command("nodeinfo", nodeinfo)
-  .command("tunnel", tunnel)
-  .command("completions", new CompletionsCommand())
-  .command("webfinger", webfinger)
-  .command("help", new HelpCommand().global());
-
-if (import.meta.main) {
-  await command.parse(Deno.args);
+async function main() {
+  const result = run(command, {
+    programName: "fedify",
+    help: "both",
+  });
+  if (result.command === "init") {
+    runInit(result);
+  }
+  if (result.command === "lookup") {
+    await runLookup(result);
+  }
+  if (result.command === "webfinger") {
+    runWebFinger(result);
+  }
+  if (result.command === "inbox") {
+    runInbox(result);
+  }
+  if (result.command === "nodeinfo") {
+    runNodeInfo(result);
+  }
+  if (result.command === "tunnel") {
+    runTunnel(result);
+  }
 }
+
+await main();
