--- packages/cli/src/table.ts
@@ -1,17 +1,0 @@
-export const tableStyle = {
-  top: "─",
-  topMid: "┬",
-  topLeft: "╭",
-  topRight: "╮",
-  bottom: "─",
-  bottomMid: "┴",
-  bottomLeft: "╰",
-  bottomRight: "╯",
-  left: "│",
-  leftMid: "├",
-  mid: "─",
-  midMid: "┼",
-  right: "│",
-  rightMid: "┤",
-  middle: "│",
-};
